Extract performance attribution contracts
