+++ src/app/services/advisor_brief_service.py
@@ -20,12 +20,12 @@ from app.contracts.advisor_brief import (
     AdvisorBriefWorkflowPackRunReviewActionRequest,
     AdvisorBriefWorkflowPackTaskFlow,
 )
+from app.contracts.performance_attribution import AttributionSummaryView
 from app.contracts.performance_contribution import (
     ContributionPositionView,
     ContributionSummaryView,
 )
 from app.contracts.performance_workspace import (
-    AttributionSummaryView,
     PerformanceComparativeSummary,
     PerformanceWorkspaceResponse,
 )
